:package: Move GQL queries into separate files
