+++ client/constants/footerLinks.js
@@ -1,0 +1,14 @@
+const links = [
+  {
+    label: "Github",
+    icon: "github",
+    url: "https://github.com/jcserv/connectu",
+  },
+  {
+    label: "Devpost",
+    icon: "devpost",
+    url: "https://devpost.com/software/connectu-q2cm8o",
+  },
+];
+
+export default links;
