--- client/data/footerLinks.js
@@ -1,19 +1,0 @@
-const links = [
-  {
-    label: "Github Repo - Frontend",
-    icon: "github",
-    url: "https://github.com/jcserv/connectu-fe",
-  },
-  {
-    label: "Github Repo - Backend",
-    icon: "github",
-    url: "https://github.com/AipioxTechson/connectu-be",
-  },
-  {
-    label: "Devpost",
-    icon: "devpost",
-    url: "https://devpost.com/software/connectu-q2cm8o",
-  },
-];
-
-export default links;
