+++ client/pages/chat/[id].jsx
@@ -1,10 +1,13 @@
-import { gql } from "@apollo/client";
 import { NextSeo } from "next-seo";
 import React from "react";
 
 import client from "../../apollo-client";
 import ChatInfo from "../../components/ChatInfo";
 import { localesArr } from "../../content/locale";
+import {
+  GET_GROUPCHAT_IDS_QUERY,
+  GET_GROUPCHAT_QUERY,
+} from "../../gql/GroupChat";
 
 export default function Chat({ chat }) {
   return (
@@ -24,13 +27,7 @@ export async function getStaticPaths() {
       getAllGroupChatIds: { groupChats },
     },
   } = await client.query({
-    query: gql`
-      query getAllGroupChatIds {
-        getAllGroupChatIds {
-          groupChats
-        }
-      }
-    `,
+    query: GET_GROUPCHAT_IDS_QUERY,
   });
   const paths =
     groupChats.length > 0 &&
@@ -51,15 +48,7 @@ export async function getStaticProps(context) {
   const {
     data: { getGroupChat },
   } = await client.query({
-    query: gql`
-      query getGroupChat($id: String!) {
-        getGroupChat(id: $id) {
-          name
-          description
-          links
-        }
-      }
-    `,
+    query: GET_GROUPCHAT_QUERY,
     variables: { id },
   });
   return {
